Move iTunesResponse struct to seperate file
